.9v5k9wt: fixed a probably long-standing bug with _average
It turns out that Chrome doesn't auto-parse, eg, "2.00" as a Number, so I have to be a bit smarter or I get compiler crashes. Probably a regression caused by the compiler getting smarter at some point.

+++ querki/scalajs/src/main/scala/org/querki/facades/raty/Raty.scala
@@ -58,8 +58,8 @@ class RatyOptionBuilder(val dict:OptMap) extends JSOptionBuilder[RatyOptions, Ra
   /**
    * Used when we want starts with a saved rating.
    */
-  def score(v:Int) = jsOpt("score", v)
-  def score(v:js.ThisFunction0[dom.Element, Int]) = jsOpt("score", v)
+  def score(v:Double) = jsOpt("score", v)
+  def score(v:js.ThisFunction0[dom.Element, Double]) = jsOpt("score", v)
   
   /**
    * Some place to display the hints or the cancelHint.
